Split out irqflags_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

We're working on only pulling in type definitions to sched.h whenever
possible.

Conflicts:
	include/linux/sched.h
(cherry picked from commit 9983deb26d9021aecd971d25abf4cd263c72c385)
